djangoend/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-extensions==3.0.9
django-seed==0.2.2
djangorestframework==3.12.1
Faker==4.1.3
ipython==5.5.0
librosa==0.8.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.1
sklearn==0.0
sklearn-pandas==1.8.0
six==1.15.0
spotipy==2.17.1
sqlparse==0.3.1
text-unidecode==1.3
